# src/main/resources/static/swagger/mng-api-docs.yaml
openapi: 3.0.3
info:
  title: 관리자 REST API 명세 ( yaml )
  description: 관리자 REST API 명세 with Swagger ( yaml )
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - Authorization: []
paths:
...
  /mng/bbs/proc:
    put:
      tags:
        - bbs-mng-controller
      summary: 게시판 수정
      operationId: updateBbs
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/BbsMngVO'
            example:

        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: string
    post:
      tags:
        - bbs-mng-controller
      summary: 게시판 등록
      operationId: createBbs
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/BbsMngVO'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: string
    delete:
      tags:
        - bbs-mng-controller
      summary: 게시판 삭제
      operationId: deleteBbs
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/BbsMngVO'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: string
  /mng/bbs/code/check:
    post:
      tags:
        - bbs-mng-controller
      summary: 게시판 코드 중복검사
      operationId: checkBbsCode
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/BbsMngVO'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: boolean
  /mng/{bbs_seq}/config/bbs/cnt:
    get:
      tags:
        - bbs-mng-controller
      summary: 게시판 존재여부 확인
      operationId: menuLinkList
      parameters:
        - name: bbs_seq
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: string
  /mng/bbs/list:
    get:
      tags:
        - bbs-mng-controller
      summary: 게시판 리스트
      operationId: bbsList
      parameters:
        - name: bbsMngVO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/BbsMngVO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: string
  /mng/bbs/info:
    get:
      tags:
        - bbs-mng-controller
      summary: 게시판 상세정보
      operationId: bbsForm
      parameters:
        - name: bbsMngVO
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/BbsMngVO'
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                type: string
...
  /api/mng/html/check/{html_seq}:
    get:
      tags:
        - Html 관리 컨트롤러
      summary: Html 존재여부 확인
      description: 메뉴 관리에서 html_seq 로 html 존재여부 확인
      operationId: menuLinkList_1
      parameters:
        - name: html_seq
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ResultVO'
...
components:
  schemas:
...
    BbsMngVO:
      type: object
      properties:
        searchCondition:
          type: string
          description: 검색 조건
        searchKeyword:
          type: string
          description: 검색어
        searchUseYn:
          type: string
          description: 검색 사용 여부
          example: Y or N
        searchOrderBy:
          type: string
          description: 검색 정렬 기준
        pageIndex:
          type: integer
          description: 현재 페이지
          format: int32
        pageUnit:
          type: integer
          description: 페이지 개수
          format: int32
        file:
          type: array
          description: 파일 배열
          items:
            type: string
            format: binary
        user_auth:
          type: string
        read_yn:
          type: string
        write_yn:
          type: string
        update_yn:
          type: string
        delete_yn:
          type: string
        category_nm:
          type: string
        category_del_yn:
          type: string
        category_order:
          type: string
        category_cd:
          type: string
        file_cnt:
          type: string
        bbs_type:
          type: string
        mobile_bbs_type:
          type: string
        bbs_dft_list_cnt:
          type: string
        bbs_page_cnt:
          type: string
        attach_file_use_yn:
          type: string
        comment_use_yn:
          type: string
        reply_use_yn:
          type: string
        notice_use_yn:
          type: string
        secret_use_yn:
          type: string
        category_use_yn:
          type: string
        pwd_use_yn:
          type: string
        date_range_use_yn:
          type: string
        file_size:
          type: string
        file_type:
          type: string
        bbs_info_seq:
          type: string
        bbs_cd:
          type: string
        bbs_nm:
          type: string
        bbs_remark:
          type: string
        bbs_use_yn:
          type: string
        bbs_del_yn:
          type: string
        inpt_seq:
          type: string
        inpt_date:
          type: string
        upd_seq:
          type: string
        upd_date:
          type: string
        date_remark:
          type: string
        bbs_cd_check_yn:
          type: string
        site_type:
          type: string
        data_range_use_yn:
          type: string
...
    ResultVO:
      type: object
      properties:
        resultMsg:
          type: string
          description: 응답 메시지
        resultCode:
          type: integer
          description: 응답 코드
          format: int32
        result:
          type: object
          additionalProperties:
            type: object
            description: 응답 결과 데이터
          description: 응답 결과 데이터
      description: REST API 응답 결과 객체
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT